refactor(api/applications): move updateDocuments logic into service file
